Proof of concept created (more details in desc)
This branch is containerizing the primary viewport that stuff gets
rendered on. It does this by creating an ImGui window and then
takes the x, y, width, height, the bounds info of the window and
projects the viewport to those coordnates. Now ImGui is still
rendered on top of the viewport, but the viewport window is
transparent. Mouse stuff is still finicky soooooo.
Now of course containerizing the primary viewport means that there's
nothing to draw on so you're gonna get weird artifacts when you
don't draw anything, so I draw a seperate viewport that fills the
screen with #222222 and follows the width and height of the window